refactoring _get_label_fields, 20X faster\!
