+++ fiftyone/server/main.py
@@ -288,23 +288,6 @@ def _numeric_bounds(view, numerics):
     return list(view.aggregate(bounds_pipeline))[0] if len(numerics) else {}
 
 
-def _get_label_fields(view):
-    label_fields = []
-
-    # @todo is this necessary?
-    label_fields.append({"_id": {"field": "_id"}})
-
-    # @todo can we remove the "_id" nesting?
-    for k, v in view.get_field_schema().items():
-        d = {"field": k}
-        if isinstance(v, fof.EmbeddedDocumentField):
-            d["cls"] = v.document_type.__name__
-
-        label_fields.append({"_id": d})
-
-    return label_fields
-
-
 def _numeric_distribution_pipelines(view, pipeline, buckets=50):
     numerics = view._dataset.get_field_schema(ftype=fof.IntField)
     numerics.update(view._dataset.get_field_schema(ftype=fof.FloatField))
